revise r16089/https://trac.osgeo.org/gdal/ticket/2781; fix makefile for plugin; remove warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@16090 f0d54148-0727-0410-94bb-9a71ac55c965

+++ frmts/georaster/georaster_wrapper.cpp
@@ -2134,7 +2134,7 @@ bool GeoRasterWrapper::UncompressDeflate( unsigned long nDestLen,
 
 bool GeoRasterWrapper::CompressDeflate( void* pData )
 {
-    unsigned long nLen = (nBlockBytes * 1.1) + 12;
+    unsigned long nLen = ((unsigned long)(nBlockBytes * 1.1)) + 12;
 
     GByte* pabyBuf = (GByte*) CPLMalloc( nLen );
